requirements.txt
Flask==1.1.2
requests==2.22.0
mysql_connector_repackaged==0.3.1
